Lifted specialization filtering memo up to ChatRoom component so that the result can be drilled down into child components. This is fed into the new disabled prop of ChatInput to grey out the input box and buttons whenever there are no specializations available due to group membership or isActive being false.
